chore(next-gen): cleanup qrcode page components
chore(next-gen): curry getNavigateToQRCode function
