+++ apps/next/src/pages/AccountManagement/components/QRCode/components/AddressSelector/index.ts
@@ -1,0 +1,1 @@
+export * from './AddressSelector';
